+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+## Fixed
+- Treasure balance is now `1m` instead of `1bn` to avoid overflow
+- Double message delivery
+- Case where multiple messages were sent to non-existent contract
+
 ## [1.1.0] - 2023-01-02
 
 ## Changed